+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Version without "v"
+- Build and upload deb + rpm packages
+
 ## [v2025-03-13-1731]
 
 ### Fixed